requirements.txt
pytest
langgraph>=0.2.0
langgraph-cli[inmem]>=0.1.0
langgraph-checkpoint-sqlite>=0.1.0
pdf2image>=1.17.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain-openai>=0.1.0
paddleocr>=2.7.0
